+++ packages/react-components/src/components/index.ts
@@ -15,9 +15,9 @@ export { SendBox } from './SendBox';
 export { _MentionPopover } from './MentionPopover';
 
 /* @conditional-compile-remove(image-overlay) */
-export { ImageOverlay } from './ImageGallery';
+export { ImageOverlay } from './ImageOverlay';
 /* @conditional-compile-remove(image-overlay) */
-export type { ImageOverlayStrings } from './ImageGallery';
+export type { ImageOverlayStrings } from './ImageOverlay';
 
 export type { SendBoxProps, SendBoxStrings, SendBoxStylesProps } from './SendBox';
 
@@ -37,7 +37,7 @@ export type {
 } from './MentionPopover';
 
 /* @conditional-compile-remove(image-overlay) */
-export type { ImageOverlayProps } from './ImageGallery';
+export type { ImageOverlayProps } from './ImageOverlay';
 
 /* @conditional-compile-remove(image-overlay) */
 export type { InlineImageOptions, InlineImage } from './ChatMessage/ChatMessageContent';
