+++ packages/react-composites/tests/browser/chat/hermetic/ImageOverlay.test.ts
