+++ packages/storybook/stories/ImageOverlay/snippets/ImageOverlay.snippet.tsx
